requirements.txt
tensorflow-macos==2.16.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
scikit-learn==1.5.0
tqdm==4.66.4
matplotlib==3.9.0
jupyterlab==4.2.3
streamlit==1.38.0
